요구사항 변경 : product 마다의 가격을 알기위해 productInfo 로 변경한다.

+++ demo/src/main/java/com/yoon/demo/vo/PurchaseOneLog.java
@@ -10,5 +10,5 @@ public class PurchaseOneLog {
     String userId; // uId-0001
     String productId; // pg-0001
     String purchasedDt; // 20230201070000
-    Long price;
+    Integer price;
 }
